requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pandas
numpy
scipy
matplotlib
openai
sentence-transformers
faiss-cpu
